pyproject.toml
[tool.poetry]
name = "test_harness"
version = "0.3.0"
description = ""
authors = ["Joshua Salzedo <jsalzedo0@saddleback.edu>"]

[tool.poetry.dependencies]
python = ">=3.8"
canvasapi = "^2.1.0"
toml = "^0.10.2"
cattrs = "^1.2.0"
attrs = "^20.3.0"
loguru = "^0.5.3"
humanfriendly = "^9.1"
click = "^7.1.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
ipython = "^7.19.0"

[tool.poetry.scripts]
gen_upload_config = "test_harness.mass_uploader.generate_config:cli"
mass_upload = "test_harness.mass_uploader.main:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
